tests/wpt/web-platform-tests/app-history/navigate-event/transitionWhile-resolve.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  appHistory.onnavigatesuccess = t.step_func_done(e => {
    assert_equals(location.hash, "#1");
    assert_equals(e.constructor, Event);
    assert_false(e.bubbles);
    assert_false(e.cancelable);
  });
  appHistory.onnavigateerror = t.step_func_done(assert_unreached);
  appHistory.onnavigate = e => {
    e.transitionWhile(new Promise(resolve => t.step_timeout(resolve, 0)));
  };

  location.href = "#1";
  assert_equals(location.hash, "#1");
}, "AppHistoryNavigateEvent's transitionWhile() should proceed if the given promise resolves");
</script>
